refactor: modify diary folder structure
