+++ src/pages/admin/diary/create/index.tsx
@@ -1,4 +1,4 @@
-import DiaryCreateComponent from 'components/Diary/components/Create'
+import DiaryCreateComponent from 'components/Diary/Admin/Create'
 import Layout from 'components/Layout/AdminLayout'
 
 const AdminDiaryCreate: React.FC = () => {
