+++ src/pages/admin/diary/index.tsx
@@ -1,4 +1,4 @@
-import AdminDiaryComponent from 'components/Diary/components/Admin'
+import AdminDiaryComponent from 'components/Diaries/Admin'
 import Layout from 'components/Layout/AdminLayout'
 
 const AdminDiaryCreate: React.FC = () => {
